+++ src/FileAdder/FileAdder.php
@@ -40,6 +40,9 @@ class FileAdder
     /** @var array */
     protected $customProperties = [];
 
+    /** @var array */
+    protected $manipulations = [];
+
     /** @var string */
     protected $pathToFile;
 
@@ -206,6 +209,20 @@ class FileAdder
     }
 
     /**
+     * Set the manipulations.
+     *
+     * @param array $manipulations
+     *
+     * @return $this
+     */
+    public function withManipulations(array $manipulations)
+    {
+        $this->manipulations = $manipulations;
+
+        return $this;
+    }
+
+    /**
      * Set properties on the model.
      *
      * @param array $properties
@@ -320,8 +337,10 @@ class FileAdder
         $media->mime_type = File::getMimetype($this->pathToFile);
         $media->size = filesize($this->pathToFile);
         $media->custom_properties = $this->customProperties;
-        $media->manipulations = [];
+
         $media->responsive_images = [];
+
+        $media->manipulations = $this->manipulations;
 
         $media->fill($this->properties);